+++ CHANGELOG.md
@@ -22,6 +22,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
     The endpoint will only use search-after when the `searchContext` parameter
     is provided in a request.
 
+### Fixed
+
+- **CUMULUS-2971**
+  - Updated `@cumulus/aws-client/S3ObjectStore` class to take string query parameters
+    and its methods `signGetObject` and `signHeadObject` to take parameter presignOptions
+
 ## [v13.2.1] 2022-8-10 [BACKPORT]
 
 ### Notable changes